Removes many gcc8 warnings other than "stringop-truncation" warnings (warnings that strncat may truncate source string in places where we INTEND strncat to truncate the source string, but which results in an unterminated destination string in the case that truncation happens).

+++ src/hashtable.c
@@ -61,7 +61,7 @@ create_hashtable(unsigned int minsize,
     h->entrycount   = 0;
     h->hashfn       = hashf;
     h->eqfn         = eqf;
-    h->loadlimit    = (unsigned int) ceil(size * max_load_factor);
+    h->loadlimit    = ceil(size * max_load_factor);
     return h;
 }
 
@@ -134,7 +134,7 @@ hashtable_expand(struct hashtable *h)
         }
     }
     h->tablelength = newsize;
-    h->loadlimit   = (unsigned int) ceil(newsize * max_load_factor);
+    h->loadlimit   = ceil(newsize * max_load_factor);
     return -1;
 }
